Fix the type of conversation.Priority to float64 #253
```
Unexpected error: %s json: cannot unmarshal number 0.0151970564935 into Go struct field Channel.priority of type int
```

+++ conversation.go
@@ -28,7 +28,7 @@ type conversation struct {
 	Unlinked           int      `json:"unlinked"`
 	NameNormalized     string   `json:"name_normalized"`
 	NumMembers         int      `json:"num_members"`
-	Priority           int      `json:"priority"`
+	Priority           float64  `json:"priority"`
 	// TODO support pending_shared
 	// TODO support previous_names
 }
